+++ CHANGES.md
@@ -2,8 +2,7 @@
 
 Released: --
 
-- Deprecate the `__version__` attribute. Use feature detection, or
-`importlib.metadata.version("apiflask")`, instead. 
+- Deprecate the `__version__` attribute. Use feature detection, or `importlib.metadata.version("apiflask")`, instead.
 
 
 ## Version 2.1.1